+++ meetup2apricot/photo_cache.py
@@ -1,5 +1,6 @@
 """Cache featured photos from Meetup events."""
 
+from . import dryrun
 import re
 import pickle
 from urllib.parse import urlparse
@@ -17,15 +18,17 @@ class PhotoCache:
         urls_to_paths,
         photo_retriever,
         cache_path,
+        dryrun=False,
     ):
         """Initialize with local and Wild Apricot directory paths, an initial
         mapping of Meetup photo URLs to Wild Apricot photo paths, a photo
-        retriever, and a path to the cache file."""
+        retriever, a path to the cache file, and a dry run flag."""
         self.local_directory = local_directory
         self.apricot_directory = apricot_directory
         self.urls_to_paths = urls_to_paths
         self.photo_retriever = photo_retriever
         self.cache_path = cache_path
+        self.dryrun = dryrun
 
     def cache_photo(self, meetup_event):
         """Cache a Meetup event's photo for copying to Wild Apricot and return
@@ -57,6 +60,12 @@ class PhotoCache:
         extension = PurePosixPath(urlparse(meetup_event.photo_url).path).suffix
         return file_name + extension
 
+    def assure_local_directory(self):
+        """Assure that the local photo directory exists."""
+        if not self.local_directory.is_dir():
+            self.local_directory.mkdir()
+
+    @dryrun.method()
     def persist(self):
         """Persist cache to a file."""
         with self.cache_path.open("wb") as f:
@@ -71,20 +80,14 @@ class PhotoCache:
         return f"{shorter_name}{date:%Y-%m-%d}"
 
 
-def make_photo_cache(cache_path, local_directory, apricot_directory, photo_retriever):
-    """Initialize with path to a file caching mapping of Meetup photo URLs to
-    Wild Apricot photo paths, local and Wild Apricot directory paths, and a
-    photo retriever."""
-    if not local_directory.is_dir():
-        local_directory.mkdir()
+def load_cached_photo_urls(cache_path):
+    """Return a mapping of Meetup photo URLs to Wild Apricot photo paths,
+    either loaded from a cache file or initialized to a default."""
     if cache_path.exists():
         with cache_path.open("rb") as f:
-            urls_to_paths = pickle.load(f)
+            return pickle.load(f)
     else:
-        urls_to_paths = {None: None}
-    return PhotoCache(
-        local_directory, apricot_directory, urls_to_paths, photo_retriever, cache_path
-    )
+        return {None: None}
 
 
 # vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4 autoindent